data_pipeline/requirements.txt
pandas>=2.0
pyarrow>=14.0
s3fs>=2024.0.0
PyYAML>=6.0
requests>=2.28
sentence-transformers>=2.2
